requirements.txt
fastapi
uvicorn
kafka-python
boto3
pydantic
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
openai
langchain
gTTS
whisper
PyPDF2
python-docx
spacy
redis
httpx
SpeechRecognition
pydub
